refactor: Update bodyParser middleware to parse application/x-www-form-urlencoded and application/json
This commit updates the bodyParser middleware in app.js to include parsing of both application/x-www-form-urlencoded and application/json request bodies. It also adds cookie parsing using the cookieParser middleware. This improves the handling of incoming request data and ensures that the server can properly parse and process form data and JSON payloads.