refactor: Update userSingUp.js to include userModel and chalk
This commit updates the userSingUp.js file to include the userModel and chalk dependencies. The userModel is imported to create a new user instance and save it to the database. The chalk library is imported to log a success message in green color when a new user is created. This change enhances the functionality of the user signup process and improves the logging experience for developers.